feat: Extend Question with IdentificationQuestion
Extend the abstract Question class with
IdentificationQuestion class and implement the
checkAnswer and getAnswers methods.

checkAnswer will first convert all the answers
in the question to lower case then checks if the
given answer is inside the List of answers.

getAnswers will just return the List of answers
in the question.

This class was added to support Identification
based questions that are not necessarily
case-sensitive.